chat_manager/requirements.txt
Django~=5.1.4
llama-cpp-python~=0.2.18
psycopg2-binary
pgvector~=0.3.6
python-decouple~=3.8
djangorestframework~=3.15.2
langchain-text-splitters~=0.3.5
langchain~=0.3.15
langchain-core~=0.3.31
langchain-community~=0.3.15
llama-index-core~=0.12.12